+++ packages/uploader-server/src/middlewares.ts
@@ -1,3 +1,4 @@
 export * from './middlewares/local-storage';
 export * from './middlewares/companion';
 export * from './middlewares/static-resource';
+export * from './middlewares/dynamic-resource';